+++ riffle-server/src/store/mem/mod.rs
@@ -17,7 +17,6 @@
 
 pub mod budget;
 pub mod buffer;
-pub mod buffer_core;
 pub mod capacity;
 pub mod ticket;
 pub use await_tree::InstrumentAwait;
