+++ riffle-server/src/store/spill/spill_test.rs
@@ -15,7 +15,6 @@ pub mod tests {
     };
     use crate::runtime::manager::RuntimeManager;
     use crate::store::hybrid::{HybridStore, PersistentStore};
-    use crate::store::mem::buffer_core::BufferOps;
     use crate::store::spill::spill_test::mock::MockStore;
     use crate::store::spill::storage_flush_handler::StorageFlushHandler;
     use crate::store::spill::storage_select_handler::StorageSelectHandler;
